Move to the new utils structure
