--- src/utils/format.js
@@ -1,4 +1,0 @@
-import { format } from 'date-fns'
-
-export const formatHtmlDatetime = date =>
-  format(date, 'YYYY-MM-DDTHH:mm:ss.SSSZ')
